Do letter colouring logic with context of full word
Previously the logic for showing the state of
letter looked only at the letter and the whole
target word, ignoring the other letters that had
been guessed.

The fix now works by looking at both words
in full, meaning that putting a second copy of
a letter that only appears once in the target
word will be be marked as incorrect/absent.